warehouse/requirements.txt
six
pyyaml
requests
python-resources
Django>=2.0.0
djangorestframework>=3.0,<4.0
opentracing>=2.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
